Avoid double-precision operations.
The hardware FPU is enabled for the CF2, but it only supports
single-precision operations.  Double-precision operations have crept
into the firmware in three ways:

1. Use of unqualified floating-point constants, e.g. "3.5" instead of
   "3.5f".  The most insidious case of this is the definition of M_PI in
   math.h.  I've qualified all constants but one, in M2T, which had a
   net negative impact by confusing the optimizer.

2. Use of the double-precision versions of standard library functions.
   In C, using "sin" over "sinf" implies promoting to double, performing
   the operation, and then converting back.  I've switched all math
   routines to the float variants.

3. Passing floats as arguments to varargs functions, like printf.  This
   is basically unavoidable in C (fixed in C++), so I haven't fixed it.

This commit reduces code size by 8% (GCC 4.8.3), by selecting simpler
library routines and enabling greater use of hardware floating point.
It will likely have a proportionally larger impact on performance
(specifically time spent in idle), but I haven't evaluated it precisely.

There is still work to be done; in particular, the current compilation
settings are preventing use of the VSQRT instruction.

+++ drivers/src/lps25h.c
@@ -236,7 +236,7 @@ bool lps25hGetData(float* pressure, float* temperature, float* asl)
 #include "math.h"
 // Constants used to determine altitude from pressure
 #define CONST_SEA_PRESSURE 102610.f //1026.1f //http://www.meteo.physik.uni-muenchen.de/dokuwiki/doku.php?id=wetter:stadt:messung
-#define CONST_PF 0.1902630958 //(1/5.25588f) Pressure factor
+#define CONST_PF 0.1902630958f //(1/5.25588f) Pressure factor
 #define CONST_PF2 44330.0f
 #define FIX_TEMP 25         // Fixed Temperature. ASL is a function of pressure and temperature, but as the temperature changes so much (blow a little towards the flie and watch it drop 5 degrees) it corrupts the ASL estimates.
                             // TLDR: Adjusting for temp changes does more harm than good.
@@ -251,7 +251,7 @@ float lps25hPressureToAltitude(float* pressure/*, float* ground_pressure, float*
     {
         //return (1.f - pow(*pressure / CONST_SEA_PRESSURE, CONST_PF)) * CONST_PF2;
         //return ((pow((1015.7 / *pressure), CONST_PF) - 1.0) * (25. + 273.15)) / 0.0065;
-        return ((pow((1015.7 / *pressure), CONST_PF) - 1.0) * (FIX_TEMP + 273.15)) / 0.0065;
+        return ((powf((1015.7f / *pressure), CONST_PF) - 1.0f) * (FIX_TEMP + 273.15f)) / 0.0065f;
     }
     else
     {
